call Lapin queue direct; move ext. queue to broker mod
