+++ src/beat/tests.rs
@@ -8,7 +8,7 @@ use async_trait::async_trait;
 use chrono::{DateTime, Utc};
 use futures::Stream;
 
-use crate::app::CeleryQueue;
+use crate::broker::Queue;
 use crate::error::{BrokerError, ProtocolError};
 use crate::{
     protocol::{Message, TryDeserializeMessage},
@@ -313,7 +313,7 @@ impl BrokerBuilder for DummyBrokerBuilder {
         unimplemented!()
     }
 
-    fn declare_queue(self, _queue: CeleryQueue) -> Self {
+    fn declare_queue(self, _queue: Queue) -> Self {
         unimplemented!()
     }
