+++ src/broker/amqp.rs
@@ -11,21 +11,21 @@ use lapin::options::{
     BasicAckOptions, BasicConsumeOptions, BasicPublishOptions, BasicQosOptions, QueueDeclareOptions,
 };
 use lapin::types::FieldTable;
-use lapin::{BasicProperties, Channel, Connection, ConnectionProperties, Queue};
+use lapin::{BasicProperties, Channel, Connection, ConnectionProperties};
 use log::debug;
 use std::collections::HashMap;
 use std::str::FromStr;
 use tokio::sync::{Mutex, RwLock};
 
 use super::{Broker, BrokerBuilder};
-use crate::app::CeleryQueue;
+use crate::broker::Queue;
 use crate::error::{BrokerError, ProtocolError};
 use crate::protocol::{Message, MessageHeaders, MessageProperties, TryDeserializeMessage};
 
 struct Config {
     broker_url: String,
     prefetch_count: u16,
-    queues: Vec<CeleryQueue>,
+    queues: Vec<Queue>,
     heartbeat: Option<u16>,
 }
 
@@ -58,7 +58,7 @@ impl BrokerBuilder for AMQPBrokerBuilder {
     }
 
     /// Declare a queue to process during broker build time.  
-    fn declare_queue(mut self, queue: CeleryQueue) -> Self {
+    fn declare_queue(mut self, queue: Queue) -> Self {
         self.config.queues.push(queue);
         self
     }
@@ -80,7 +80,7 @@ impl BrokerBuilder for AMQPBrokerBuilder {
         let consume_channel = conn.create_channel().await?;
         let produce_channel = conn.create_channel().await?;
        
-        let mut queues: HashMap<String, Queue> = HashMap::new();
+        let mut queues: HashMap<String, lapin::Queue> = HashMap::new();
         let mut queue_options: HashMap<String, QueueDeclareOptions> = HashMap::new();
   
         for queue in &self.config.queues { 
@@ -150,7 +150,7 @@ pub struct AMQPBroker {
     /// Mapping of queue name to Queue struct.
     ///
     /// This is only wrapped in RwLock for interior mutability.
-    queues: RwLock<HashMap<String, Queue>>,
+    queues: RwLock<HashMap<String, lapin::Queue>>,
 
     queue_declare_options: HashMap<String, QueueDeclareOptions>,
